Use size_t for Totem's get_runtime_type_id

+++ include/GM/Framework/Totem/Component.h
@@ -26,7 +26,7 @@ public:
 	Component(const EntityPtr &owner, const std::string &name);
 	virtual ~Component();
 
-	unsigned int get_runtime_type_id() const override;
+	size_t get_runtime_type_id() const override;
 	const std::string &get_name() const override;
 	virtual const Entity *get_owner() const override;
 	virtual Entity *get_owner() override;
